A22DAMPIMOBI-2098 : [Điểm danh] sửa lại theo ghép api điểm danh khuôn mặt

+++ lib/diem_danh_module/domain/model/thong_ke_diem_danh_ca_nhan_model.dart
@@ -1,7 +1,7 @@
 class ThongKeDiemDanhCaNhanModel {
   int? soLanDiMuon;
   int? soLanVeSom;
-  int? soNgayLamViec;
+  double? soNgayLamViec;
   int? soLanChamCongThuCong;
   int? soNgayVangMatKhongLyDo;
   int? soNgayNghiCoLyDo;
